+++ NEWS.md
@@ -2,6 +2,8 @@
 - Fix error when writting and reading data warehouse (#298)
 - Fix bugs when reading data from data warehouse (#300)
 - Upgrade package version dependency.
+- Unknown municipality codes but known county codes will be recorded to `xx99`
+  in `geo_recode()` function. (#302)
 
 # orgdata 1.2.0
 - Function `read_file()` accept SPSS file too (#297)